+++ src/main.c
@@ -92,6 +92,7 @@ int main(int argc, char** argv) {
 
     /* colors */
     start_color();
+
     init_pair(1, COLOR_BLUE, COLOR_BLACK);
     init_pair(2, COLOR_GREEN, COLOR_BLACK);
     init_pair(3, COLOR_YELLOW, COLOR_BLACK);